requirements.txt
langchain
langchain-community
langchain-google-genai
fastapi
uvicorn
python-dotenv
pymupdf
faiss-cpu